requirements.txt
# requirements.txt
flask==2.0.1
flask-socketio==5.1.1
paho-mqtt==1.6.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
plotly==5.3.1
dash==2.0.0
python-dotenv==0.19.0
gunicorn